+++ cookbooks/mylab_dhcp/metadata.rb
@@ -1,25 +1,25 @@
-name 'maxlab_dhcp'
+name 'mylab_dhcp'
 maintainer 'Maxwell Spangler'
 maintainer_email 'maxcode@maxwellspangler.com'
 license 'All Rights Reserved'
-description 'Deploy a DHCP server on Red Hat Linux systems'
-long_description 'Deploy a DHCP server on Red Hat Linux systems'
-version '1.2.0'
-chef_version '>= 13.0'
+description 'Wrapper for maxlab_dhcp, experimentation'
+version '1.0.0'
+chef_version '>= 14.0'
 
 # The `issues_url` points to the location where issues for this cookbook are
 # tracked.  A `View Issues` link will be displayed on this cookbook's page when
 # uploaded to a Supermarket.
 #
-# issues_url 'https://github.com/<insert_org_here>/maxlab_dhcp/issues'
+# issues_url 'https://github.com/<insert_org_here>/mylab_dhcp/issues'
 
 # The `source_url` points to the development repository for this cookbook.  A
 # `View Source` link will be displayed on this cookbook's page when uploaded to
 # a Supermarket.
 #
-# source_url 'https://github.com/<insert_org_here>/maxlab_dhcp'
+# source_url 'https://github.com/<insert_org_here>/mylab_dhcp'
 
 supports 'redhat', '>= 7.0'
 supports 'centos', '>= 7.0'
 
+depends 'maxlab_dhcp'
 depends 'maxlab_firewall'