+++ cookbooks/mylab_dhcp/recipes/default.rb
@@ -1,0 +1,12 @@
+#
+# Cookbook:: mylab_dhcp
+# Recipe:: default
+#
+# Copyright:: 2020, The Authors, All Rights Reserved.
+
+
+=begin
+#<
+This recipe performs no actions.
+#>
+=end