+++ cookbooks/mylab_dhcp/recipes/deploy.rb
@@ -1,0 +1,14 @@
+#
+# Cookbook:: mylab_dhcp
+# Recipe:: deploy
+#
+# Copyright:: 2020, The Authors, All Rights Reserved.
+
+
+=begin
+#<
+Wrapper recipe for maxlab_dhcp::deploy
+#>
+=end
+
+include_recipe 'maxlab_dhcp::deploy'